feat: update recyclebin style
