+++ src/components/cards/archivingCard/ArchivingCard.tsx
@@ -142,7 +142,9 @@ export const ArchivingCard = ({ item, isMine, isRecycle }: ArchivingCardProps) =
             {title}
           </Title>
           <Day>{createdAt}</Day>
-          {isMine ? (
+          {isRecycle ? (
+            <></>
+          ) : isMine ? (
             <>
               <PopupContainer>
                 <Popup menuList={popupMenuList} />
