+++ src/screens/recycleBin/RecycleBin.tsx
@@ -1,17 +1,16 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 
-import { useNavigation } from '@react-navigation/native'
 import LinearGradient from 'react-native-linear-gradient'
 import { useMutation, useQuery, useQueryClient } from 'react-query'
 import { useRecoilState } from 'recoil'
 
 import { deleteRecycles, getRecycles, patchRecycles } from '@/apis/recycle'
+import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
 import TwoButtonDialog from '@/components/dialogs/twoButtonDialog/TwoButtonDialog'
 import EmptyItem from '@/components/emptyItem/EmptyItem'
 import { LeftButtonHeader } from '@/components/headers/leftButtonHeader/LeftButtonHeader'
 import i18n from '@/locales'
 import { RecyclesResponse } from '@/models/Recycle'
-import { MainNavigationProp } from '@/navigations/MainNavigator'
 import { CheckArchivingState, CheckContentState } from '@/state/CheckState'
 import { colors } from '@/styles/colors'
 
@@ -21,7 +20,6 @@ import {
   BottomButtonText,
   BottomButtonTitle,
   Container,
-  TabContainer,
 } from './RecycleBin.style'
 import { RecycleBinTab } from './tabs/RecycleBinTab'
 
@@ -29,7 +27,6 @@ import { RecycleBinTab } from './tabs/RecycleBinTab'
  * 마이페이지 '휴지통'
  */
 export const RecycleBin = () => {
-  const navigation = useNavigation<MainNavigationProp>()
   const queryClient = useQueryClient()
 
   const [editMode, setEditMode] = useState(false)
@@ -60,21 +57,6 @@ export const RecycleBin = () => {
       setIsCheckArchiving([])
       setIsCheckContent([])
     },
-  })
-
-  useEffect(() => {
-    navigation.setOptions({
-      /**
-       * header
-       */
-      header: () => (
-        <LeftButtonHeader
-          title={i18n.t('recycleBin')}
-          rightButtonText={editMode ? i18n.t('complete') : i18n.t('edit')}
-          rightButtonClick={handleEditMode}
-        />
-      ),
-    })
   })
 
   /**
@@ -114,65 +96,65 @@ export const RecycleBin = () => {
   }
 
   return (
-    <Container>
-      {recycleData ? (
-        <>
-          <TabContainer>
-            <RecycleBinTab
-              contents={recycleData.contents}
-              archivings={recycleData.archivings}
-              editMode={editMode}
-            />
-          </TabContainer>
-
-          {editMode && (
-            <BottomButtonContainer>
-              <LinearGradient
-                style={{
-                  width: '100%',
-                  height: '100%',
-                  flexDirection: 'row',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                  padding: 24,
-                }}
-                colors={[colors.white, colors.gray500]}
-              >
-                <BottomButton onPress={handleDeleteDialog}>
-                  <BottomButtonText>{i18n.t('allDelete')}</BottomButtonText>
-                </BottomButton>
-                {isCheckArchiving.length > 0 || isCheckContent.length > 0 ? (
-                  <>
-                    <BottomButtonTitle>
-                      {i18n.t('numberOfSelectItem', {
-                        number: isCheckArchiving.length + isCheckContent.length,
-                      })}
-                    </BottomButtonTitle>
-                  </>
-                ) : (
-                  <>
-                    <BottomButtonTitle>{i18n.t('selectItem')}</BottomButtonTitle>
-                  </>
-                )}
-
-                <BottomButton onPress={handleRestore}>
-                  <BottomButtonText>{i18n.t('allRestore')}</BottomButtonText>
-                </BottomButton>
-              </LinearGradient>
-            </BottomButtonContainer>
-          )}
-
-          <TwoButtonDialog
-            isVisible={isDeleteDialogVisible}
-            title="persistentDeleteWarning"
-            completeText={i18n.t('delete')}
-            onCancel={() => setIsDeleteDialogVisible(false)}
-            onComplete={handleDelete}
+    <DefaultContainer>
+      <LeftButtonHeader
+        title={i18n.t('recycleBin')}
+        rightButtonText={editMode ? i18n.t('complete') : i18n.t('edit')}
+        rightButtonClick={handleEditMode}
+      />
+      <Container>
+        {recycleData ? (
+          <RecycleBinTab
+            contents={recycleData.contents}
+            archivings={recycleData.archivings}
+            editMode={editMode}
           />
-        </>
-      ) : (
-        <EmptyItem textKey="emptyRecycleBin" />
+        ) : (
+          <EmptyItem textKey="emptyRecycleBin" />
+        )}
+      </Container>
+      {editMode && (
+        <BottomButtonContainer>
+          <LinearGradient
+            style={{
+              width: '100%',
+              height: '100%',
+              flexDirection: 'row',
+              justifyContent: 'space-between',
+              alignItems: 'center',
+              padding: 24,
+            }}
+            colors={[colors.white, colors.gray500]}
+          >
+            <BottomButton onPress={handleDeleteDialog}>
+              <BottomButtonText>{i18n.t('allDelete')}</BottomButtonText>
+            </BottomButton>
+            {isCheckArchiving.length > 0 || isCheckContent.length > 0 ? (
+              <>
+                <BottomButtonTitle>
+                  {i18n.t('numberOfSelectItem', {
+                    number: isCheckArchiving.length + isCheckContent.length,
+                  })}
+                </BottomButtonTitle>
+              </>
+            ) : (
+              <>
+                <BottomButtonTitle>{i18n.t('selectItem')}</BottomButtonTitle>
+              </>
+            )}
+            <BottomButton onPress={handleRestore}>
+              <BottomButtonText>{i18n.t('allRestore')}</BottomButtonText>
+            </BottomButton>
+          </LinearGradient>
+        </BottomButtonContainer>
       )}
-    </Container>
+      <TwoButtonDialog
+        isVisible={isDeleteDialogVisible}
+        title="persistentDeleteWarning"
+        completeText={i18n.t('delete')}
+        onCancel={() => setIsDeleteDialogVisible(false)}
+        onComplete={handleDelete}
+      />
+    </DefaultContainer>
   )
 }
