+++ src/screens/recycleBin/tabs/AllTab.tsx
@@ -12,16 +12,16 @@ import { SimpleContent } from '@/models/SimpleContent'
 import { CheckArchivingState, CheckContentState } from '@/state/CheckState'
 
 import {
-  CheckBox,
   Container,
+  TabItemContainer,
+  SearchDataText,
+  CheckBox,
   ContentListContainer,
   GrayDivider,
-  SearchDataText,
   TabItemCardContainer,
-  TabItemContainer,
   Title,
   YellowCheck,
-} from '../RecycleBin.style'
+} from './Tab.style'
 
 /**
  * 전체 탭
@@ -59,7 +59,7 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
    */
   const renderItem: ListRenderItem<SimpleContent> = ({ item }) => {
     return (
-      <View>
+      <>
         <ContentCard
           contentId={item.contentId}
           contentTitle={item.contentTitle}
@@ -77,7 +77,7 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
             <Image source={defaultIcons.yellowCheck} />
           </YellowCheck>
         )}
-      </View>
+      </>
     )
   }
 
@@ -93,26 +93,24 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
           </SearchDataText>
           <Title>{i18n.t('archiving')}</Title>
           <TabItemCardContainer>
-            {archivings !== undefined &&
+            {archivings &&
               archivings.map((item) => (
-                <>
-                  <View>
-                    <ArchivingCard
-                      key={item.archivingId}
-                      item={item}
-                      isMine={true}
-                      isRecycle={true}
-                    />
-                    {editMode && (
-                      <CheckBox onPress={() => handleCheck(item.archivingId, 'archiving')} />
-                    )}
-                    {editMode && isArchivingCheck.includes(item.archivingId) && (
-                      <YellowCheck onPress={() => handleCheck(item.archivingId, 'archiving')}>
-                        <Image source={defaultIcons.yellowCheck} />
-                      </YellowCheck>
-                    )}
-                  </View>
-                </>
+                <View key={item.archivingId}>
+                  <ArchivingCard
+                    key={item.archivingId}
+                    item={item}
+                    isMine={true}
+                    isRecycle={true}
+                  />
+                  {editMode && (
+                    <CheckBox onPress={() => handleCheck(item.archivingId, 'archiving')} />
+                  )}
+                  {editMode && isArchivingCheck.includes(item.archivingId) && (
+                    <YellowCheck onPress={() => handleCheck(item.archivingId, 'archiving')}>
+                      <Image source={defaultIcons.yellowCheck} />
+                    </YellowCheck>
+                  )}
+                </View>
               ))}
           </TabItemCardContainer>
         </TabItemContainer>
@@ -123,7 +121,7 @@ export const AllTab = ({ contents, archivings, editMode }: RecycleBinTabProps) =
           </SearchDataText>
           <Title>{i18n.t('contents')}</Title>
           <TabItemCardContainer>
-            {contents !== undefined && (
+            {contents && (
               <ContentListContainer
                 scrollEnabled={false}
                 data={contents}
